Input subsystem refactoring; see log for details
  - nom::SDL_App
* Remove SDL_App::enableKeyRepeat method; SDL_EnableKeyRepeat function has been
phased out entirely SDL2's new event subsystem, and is now taken care of for us
automagically.

  - nom::Input
* Introduce Drag'N'Drop event support via SDL2 API

* Implement preliminary support for the new SDL_WINDOWEVENT_CLOSE event; this
fixes the bug Ivan noticed that results in no action being taken on a window close event.

* Add an argument for passing along the new event struct data 'windowID' along
with several existing events: keyboard & mouse (motion, buttons). This opens the
doorway for mapping an event with a particular window.

* Implement changes to support SDL2 mouse wheel events.

* Add preprocessor defines for enabling sections of the input subsystem (for
debugging aid).

  - examples/app
* Fix the window icon not being set for all three windows under the Windows
platform.

* Fix the window title display text; show the actual window_id instead of the
index count of the array we use for holding the windows.