Refactoring of nom::SDL_App, nom::Window & examples/app
  - nom::SDL_App
* Remove isFullscreen & setFullScreen methods; remove fullscreen
instance variable.

  - nom::Window
* Re-implement the fullscreen flags previously found in SDL_App; this is
done out of necessity, since we are not limited by a single window
anymore.

* Introduce toggle_fullscreen method
* Introduce size getter method
* Introduce draw method

* Fix the setting & caching of our window_id_ & window_display_id_
variables.

* Introduce pixels and save_screenshot methods; these are not currently
working and notes have been made as to why along with commenting said
methods out.

  - examples/app
* Use the new nom::Window fullscreen methods and implement checking for
window_id so we have working fullscreen toggle that is based on the
current window that has focus.

* I also introduce using set_viewport here so that when fullscreen has
been activated, it upscales the contents of the window for us.