JsonSerializer: Major refactoring -- new features, too!
* Introduce support for JSON mapped objects -- yay! All we have left
now is mostly clean up work in this class.

* Use JsonCpp interface directly within serialize & unserialize methods;
this allows us to remove the redundant JsonCppSerializer class entirely.

* Use JsonCpp's interface instead of JsonCppValue for internal handling
of JSON serialization/unserialization. (JsonCppValue is now officially
marked deprecated).

* Move const JSONCPP_INDENTION_LEVEL & Feature enum from
JsonCppSerializer class to JsonSerializer.

* Introduce dump interface, deriving from nom::Value's implementation.

* Re-inherit from IJsonSerializer. I plan on *hopefully* designing an
interface that allows for swapping this parser class out for another,
in case of the need arising down the road.