travis-ci: Fix a few build issues
* Fix tar extraction err when using BSD tar by using single quotes
in the web location passed to wget.

* Fix CMake build warning regarding the missing OSX 10.7 SDK by
providing a local copy for the build to use; the SDK tar ball is
hosted from my DropBox account.

* Fix the installation path for unit test resources when using
XCode project files (generated by CMake); this resolves missing
resources that are needed by a few of the tests.

* Add test failure messages for FontCacheTest::GlobalSystemFontsInterface.

Unfortunately, none of these fixes brings me much closer to resolving
the failing tests -- all I know is that it doesn't like something to do
with BitmapFont objects (TrueType font tests are not failing).