FileInterfaceTest: Fix all known test errs under OS X & WinOS
* Path: Fix mis-initialization of path delimiter when compiled under
Windows OS; the bug fix is changing the delimiter from "\/" to "\\".
  This was causing trouble all over the place -- working dir path
resolution (contributing to the break of RenderWindowTest::SetIcon),
the creation & removal of directory paths (breaking GUI layout tests,
file tests and so on).

* WinFile: Fix ::mkdir && ::rmdir methods; we were not handling the
return value of the creation & removal functions, and thus receiving
false positives. This helps correct the running of the tests on
Windows.

* Change the dir path of the temp dir paths that are created & removed to
/tmp/nomlib (from ~/Documents/nomlib).

* Implement proper test methodology by cleaning up (removing) the dir
paths at the time of test tear down; this helps keep the tests from
failing by simply finding its existing directories from previous test
runs, along with a reduction of code duplication...