SDL2Logger: Introduce control over auto-spacing logging messages
* Introduce NOM_LOG_CATEGORY_MEMORY, NOM_LOG_CATEGORY_MEMORY_TOTALS
enumerations; these will eventually be used for logging memory allocations
and deallocations once we finish implementing a global new/delete method
for use.

* SDL2Logger: Introduce boolean auto_space class var & ctor argument for
allowing us to optionally control whether or not to insert a space char
at the end of a log message (applicable when varadic args are used).
  The NOM_LOG_MESSAGE helper macro has been updated with the
default arg of disabling the insertion of a space char.
(This may change back in the near future...)

* Image: Update ctor & dtor trace logs to use SDL2Logger interface.
* Texture: Update ctor & dtor trace logs to use SDL2Logger interface.