Massive clean up of header files (speed up build)
* UnitTest: Move initialization of VisualUnitTest's event listener
(HTML results output) to VisualUnitTest.
  This allows us to properly separate the UnitTest sources from
VisualUnitTest sources as per intended design.

* Introduce necessary logic in the build script to let me handle
optionally building the ScaleX and HQX algorithms without
consequence (to Texture).
  When NOM_BUILD_EXTRA_RESCALE_ALGO_UNIT is enabled, NOM_USE_SCALEX
and NOM_USE_HQX defines are used within the Texture class to somewhat
gracefully handle the condition aborting the attempt and informing the
end-user that the algorithm has not been built into the engine.