Fix a few issues in Sprite interface
Sprite: Introduce virtual set_position method; this resolves the
long-standing desire to get rid of manually calling ::update.

SpriteBatch: Set the sprite's size to match the current frame in
use; this resolves the long-standing bug involving incorrect size
output.

* Sprite: Remove copy assignment operator; this isn't necessary
because we do not have pointers or custom resource needs.
* Sprite: Remove size method.
* Sprite: Re-declare update under protected scope.

* SpriteBatch: Remove copy assignment operator; this isn't necessary
because we do not have pointers or custom resource needs.
* SpriteBatch: Add virtual keyword to update method.