Fix event update logic in ListBox && QuestionDialogBox
Internal updates to the storage model of said widgets need to occur
before emitting the event (for end-user to respond to), otherwise the
data will always be inaccurate (one-off).

* QuestionDialogBox: Disable calling of base class for events; this
corrects a bug in TTcards (ContinueMenuState) when clicking on either
the title or message, due to duplicate indexes being used in both
widgets (TTcards depends on the values for cursor action).

* ListBoxLayoutTest: Re-enable event callbacks -- MOUSE_DOWN, KEY_DOWN,
MOUSE_WHEEL (debugging aid).