Fix event callbacks crash by using deferred states management
The transition of the state must be deferred until the next frame update
in order for engine callback systems, i.e.: nom::InputMapper and
nom::UIWidget, to be used within application-level code (TTcards)
that expects to be able to change its state from within said callback
systems. If not deferred, the result is invalid memory access due to
the deletion of the state before the callback is completed.

* Add IState to public headers (system.hpp).
* IState: Forward declare RenderWindow.
* IState: Rename UniquePtr typedef to unique_ptr.
* IState: Rename StateFlags enum to Flags, remove Null enum value
and clean up comments.

* IState: Introduce Type enumeration, getter & ctor argument; describe
placement of state (internal use within deferred state management).

* StateMachine: Forward declare Event, IState, RenderWindow & clean up
header files.
* StateMachine: Remove push_state method; obsolete by deferred states.
* StateMachine: Remove default parameter from data argument in methods.
* StateMachine: Remove pop_state_resume; obsolete by deferred states.
* StateMachine: Introduce next_state method.