Gradient: Fix margins bug recently introduced (tex cache)
I had forgotten to account for local margins, and thus the
relative to absolute coordinates became out of sync. The red border
color we set in case of this kind of err really paid off for
helping visualize this bug fix!

* Use resource search paths cfg for GradientTest.
* Use a different color for secondary gradient test
object to help make it stand out.

* Introduce Margins test; uses an emulation of the bug sample
caught from TTcards to verify fix.