SpriteSheet: Fix mis-matched sorting issue
This is a major redesign of the class, aimed at solving the sorting
mismatch issue that we've had problems with in the past by using a
std::map instead of std::vector. When looking up sprite frame
indexes, we expect the returned value to match the frame ID,
whereas the previous implementation was actually giving us the
element's position in the std::vector instead, which is not always
going to be equal.
  The file format spec has been altered dramatically, and invalidates
all previous versions. No backwards compatibility is provided; all
examples/tests have been updated to use the new file format spec.

* This implementation does away with the serialization feature of
nom::SpriteSheet entirely, and instead relies on the wonderful
TexturePacker software to provide this facility. Ultimately simplifying
our design and saves me time.