Overhaul of fonts system
The nom::SystemFonts and nom::PlatformSettings interfaces have been
removed, due to their greatly reduced usefulness in light of the
GUI back-end changes (from in-house to libRocket).

* Text: Remove raw_ptr, shared_ptr, unique_ptr and font_type typedefs.
* Text: Remove ::get.
* Text: The default-initialized alignment is now Alignment::NONE.

* Font: Remove raw_ptr, shared_ptr, unique_ptr and value_type typedefs.
* Font: Removed copy-on-write implementation; this should be done with
the font's clone method.

* ResourceCache: Change check for valid function to nullptr; this
ensures that we can use the interface when the resource type does
not implement a valid method.
  Additionally, the resource handler nom::create_font has been removed
(commented out, the files still exist for future reference).

* FontCacheTest: Bit of code clean up -- disabled debug logging
levels -- renamed test descriptions, etc in addition to removal of
nom::SystmeFonts and nom::Font copy-on-write features.

* examples/fonts.cpp: Refactored to not use SystemFonts feature.