WinFile: Fix returned path when a POSIX path is passed
This bug fix resolves working directory issues that arise when trying
to run unit tests from the generated project file by accepting both
POSIX and Windows delimiters as valid.

This was due to the fact that we use argv[0] in nom::init to obtain the
current working directory before switching to it, of which value gives
us a POSIX path, whereas we were expecting a native path.