Text, Gradient: Decrease # of tex allocs
This is a light optimization in light of the poor performance in
recent builds of TTcards.

* Text, Gradient: Only update the texture cache when the size
dimensions differ.

* Text, Gradient: Change err logging category from
NOM_LOG_CATEGORY_RENDER to NOM_LOG_CATEGORY_APPLICATION.

* DecoratorFinalFantasyFrame: Fix render loop; this method was
updating our custom decorator on every frame, instead of only when
the element's coordinates changed, due to lack of equality between
floating-point and integer numbers -- woops!

* UIMessageBox: Update the message text only when the string differs
from the existing.