Fix building engine with Xcode && MSVCPP generated proj files
The installation of our examples, tests, resources were not going to
the right directories for either Debug or Release targets. The output
directories should now be set correctly for both generators for both
target types (CMAKE_*_OUTPUT_DIRECTORY magic).
  Note that you must keep Debug and Release build targets in separate
trees!!! This was the major GOTCHA for me.

* Introduce several new helper shell scripts in the bin dir for
building with Xcode && msbuild.

* macros.cmake: Remove workaround code for the Windows platform now
that we have the output directories finally figured out.
* nomlib.sublime-project: Add build-release to file exclusion list.
* README.md: Update build notes with important reminder regarding
multiple build targets.