Value: Fix bug in ::erase method
This was throwing iterator debugging assertions when build with MSVCPP
during the 'PropertyTreeTest.ValueFindEraseInterface' test on
windev.local. The underlying cause for this was because we were
erasing the value and then trying to return it -- we had intended to
make a copy of it for returning after erasing it.