Value: Fix memory leaks
This was a long time coming, and feels good to finally check off the
TODO list! In short, the issues boiled down to duplicating strings
(too many) and shallow-copy instances of object values (when we needed
deep-copy instancing).

* Value: Properly free value_.object_ inside destructor. The problem
here boiled down to us forgetting to create a deep-copy instance of the
object values upon copy construction (instead just copying the pointer).
Oops!

* Value: Fix copy construction; string values and object values are now
properly deep-copied. We were dumb and were effectively duplicating a
duplicate of our strings each time. Oops!

* Value: Fix ::get_cstring && ::get_string methods; we were always
returning duplicates of the stored strings, effectively duplicating
twice. Oops!

* Value: Clean up operator = && ::swap methods.
* Value: Add const versions of operator[](key).
* Value: operator[](const char* key) has been rewritten for readability
sake. Jesus christ, was the code horrid to try debugging as-is!

* VString: This class has been rewritten to use std::string instead of
const char*, vastly simplifying memory management for nom::Value; this
is where most of the memory-related headaches stem from and alone
resolves most of the priv::duplicate_string related memory leaks.

* JsonCppSerializerTest: Update
SerializeArrayValues_ObjectInsideArrayValues &&
SerializeObjectValues_ObjectInsideArrayValues tests; these required
minor changes due to the internal fix in the copy constructor of
nom::Value -- since a deep-copy is now done upon copy, instead of a
shallow-copy, the assignment I was doing before becomes invalid. (It
should have never worked to begin with!)
  Add readability debugging aid to ::expected_out; when a test fails,
it dumps the expected output in the compressed, one-line JSON style,
which is not fun for debugging from, so this (optional) feature helps
to resolve that by saving the expected && actual JSON output in
"human friendly" files.

* JsonCppSerializerTest: Mark ExtraIO test as Disabled, instead of
commented out; this will hopefully help remind us to someday fix this
test! :-P

* JsonCppSerializerTest: Remove ::pp_out, ::pp_outs methods; these were
unused.