EventHandler: Improve accuracy of event queue stats
The maximum events count was slightly off due to the point in which
it was being calculated, and is now done in ::push_event instead of
::pop_event.

* EventHandler: Additional debugging output to see enqueued event data
is available under NOM_LOG_CATEGORY_EVENT.

* EventHandler: Fix order of event processing in ::process_events; we
were previously processing events before retrieving them, instead of
retrieving the event and then processing it. Additionally, we now check
for event retrieval errors as well as only processing an event when
there are one or more events (not unconditionally as before).

* EventHandler: The current number of enqueued events is now explicitly
tracked.

* examples/events: Remove event debugging output calls; the additional
info made available from nom::EventHandler supercedes it.