Clean up audio subsystem
This resolves all but one err message in ALAudioTest. In addition, audio
volume levels are now normalized upon setting and de-normalized upon
getting (effects Listener and SoundSource).

* Listener: Rename functions.
* Listener: Add max_volume, min_volume functions.
* Fix Listener::set_direction function by adding two Point3f arguments;
this resolves OpenAL errs related to Listener when running ALAudioTest.

* NullSoundSource: ::volume now returns 100.0f, was 0.0f; this matches
the OpenAL implementation.
* NullSoundBuffer: ::load now returns true.

* AudioDevice: Add err checks.

* Remove ::fadeOut functions; the logic of the function now exists
solely in the audio example (and is due to be replaced eventually by
the actions subsystem).

* audio example now accepts command line arguments for changing the
input audio file and testing the playback out with both the null and
music interfaces.